server/requirements.txt
pycapnp==1.2.1
